Address PR feedback, reorganize modules, add insecure provider
