+++ confmap/provider/internal/configurablehttpprovider/provider_test.go
@@ -126,7 +126,7 @@ func generateCertificate(hostname string) (cert string, key string, err error) {
 }
 
 func TestFunctionalityDownloadFileHTTP(t *testing.T) {
-	fp := newConfigurableHTTPProvider(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(answerGet))
 	defer ts.Close()
 	_, err := fp.Retrieve(context.Background(), ts.URL, nil)
@@ -214,7 +214,7 @@ func TestFunctionalityDownloadFileHTTPS(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.testName, func(t *testing.T) {
-			fp := newConfigurableHTTPProvider(HTTPSScheme, confmaptest.NewNopProviderSettings())
+			fp := newConfigurableHTTPProvider(EECScheme, confmaptest.NewNopProviderSettings())
 			// Parse url of the test server to get the port number.
 			tsURL, err := url.Parse(ts.URL)
 			require.NoError(t, err)
@@ -233,19 +233,19 @@ func TestFunctionalityDownloadFileHTTPS(t *testing.T) {
 }
 
 func TestUnsupportedScheme(t *testing.T) {
-	fp := New(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
 	_, err := fp.Retrieve(context.Background(), "https://...", nil)
 	assert.Error(t, err)
 	assert.NoError(t, fp.Shutdown(context.Background()))
 
-	fp = New(HTTPSScheme, confmaptest.NewNopProviderSettings())
+	fp = New(EECScheme, confmaptest.NewNopProviderSettings())
 	_, err = fp.Retrieve(context.Background(), "http://...", nil)
 	assert.Error(t, err)
 	assert.NoError(t, fp.Shutdown(context.Background()))
 }
 
 func TestEmptyURI(t *testing.T) {
-	fp := New(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusBadRequest)
 	}))
@@ -256,7 +256,7 @@ func TestEmptyURI(t *testing.T) {
 }
 
 func TestRetrieveFromShutdownServer(t *testing.T) {
-	fp := New(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(func(http.ResponseWriter, *http.Request) {}))
 	ts.Close()
 	_, err := fp.Retrieve(context.Background(), ts.URL, nil)
@@ -265,7 +265,7 @@ func TestRetrieveFromShutdownServer(t *testing.T) {
 }
 
 func TestNonExistent(t *testing.T) {
-	fp := New(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusNotFound)
 	}))
@@ -276,7 +276,7 @@ func TestNonExistent(t *testing.T) {
 }
 
 func TestInvalidYAML(t *testing.T) {
-	fp := New(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
 		_, err := w.Write([]byte("wrong : ["))
@@ -291,13 +291,13 @@ func TestInvalidYAML(t *testing.T) {
 }
 
 func TestScheme(t *testing.T) {
-	fp := New(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
 	assert.Equal(t, "http", fp.Scheme())
 	require.NoError(t, fp.Shutdown(context.Background()))
 }
 
 func TestValidateProviderScheme(t *testing.T) {
-	assert.NoError(t, confmaptest.ValidateProviderScheme(New(HTTPScheme, confmaptest.NewNopProviderSettings())))
+	assert.NoError(t, confmaptest.ValidateProviderScheme(New(EECIScheme, confmaptest.NewNopProviderSettings())))
 }
 
 func TestInvalidTransport(t *testing.T) {
@@ -330,7 +330,7 @@ func TestNoReloadIfConfigIsSame(t *testing.T) {
 	watcherFunc := func(_ *confmap.ChangeEvent) {
 		called.Store(true)
 	}
-	fp := newConfigurableHTTPProvider(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(answerWithCount))
 	defer ts.Close()
 	uri, err := url.Parse(ts.URL)
@@ -375,7 +375,7 @@ func TestReloadIfConfigChanges(t *testing.T) {
 	watcherFunc := func(_ *confmap.ChangeEvent) {
 		called.Store(true)
 	}
-	fp := newConfigurableHTTPProvider(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(answerWithCount))
 	defer ts.Close()
 	uri, err := url.Parse(ts.URL)
@@ -426,7 +426,7 @@ func TestContinuesRetryingOnRefreshError(t *testing.T) {
 	watcherFunc := func(_ *confmap.ChangeEvent) {
 		called.Store(true)
 	}
-	fp := newConfigurableHTTPProvider(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(answerWithCount))
 	defer ts.Close()
 	uri, err := url.Parse(ts.URL)
@@ -479,7 +479,7 @@ func TestFragmentConfiguration(t *testing.T) {
 	watcherFunc := func(_ *confmap.ChangeEvent) {
 		called.Store(true)
 	}
-	fp := newConfigurableHTTPProvider(HTTPScheme, confmaptest.NewNopProviderSettings())
+	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
 	ts := httptest.NewServer(http.HandlerFunc(answerWithConfig))
 	defer ts.Close()
 	uri, err := url.Parse(ts.URL)
