+++ confmap/provider/internal/configurablehttpprovider/testdata/otel-config-updated.yaml
