+++ confmap/provider/internal/configurablehttpprovider/testdata/otel-config.yaml
